Added SQL database via upload
mqSQL Dump - includes all the tables in the database. Only Recipes, Ingredients and Methods table has got data as prescribed in the assignment. Also, added all the query files that I used to create and insert data. Test_scripts is what I used to check the data in the tables.